Move cross-project configuration to a Gradle convention plugin
