--- build-logic/src/main/kotlin/odeon/tasks/odeon-application.gradle.kts
@@ -1,33 +1,0 @@
-/*
- * Copyright 2022 Thibault Seisel
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import com.android.build.api.variant.AndroidComponentsExtension
-import odeon.tasks.RefreshMediaStore
-
-plugins {
-    id("com.android.application") apply false
-}
-
-pluginManager.withPlugin("com.android.application") {
-    val androidComponents = extensions.getByType(AndroidComponentsExtension::class)
-
-    tasks.register<RefreshMediaStore>("refreshMediaStore") {
-        group = "emulator"
-        description = "Scan the connected device's Music folder for files to refresh MediaStore."
-        @Suppress("UnstableApiUsage")
-        adbPath.set(androidComponents.sdkComponents.adb)
-    }
-}
