+++ ui-cleanup/build.gradle.kts
@@ -15,6 +15,7 @@
  */
 
 plugins {
+    id("odeon-convention")
     id("com.android.library")
     kotlin("android")
     kotlin("kapt")
@@ -22,10 +23,6 @@ plugins {
 }
 
 android {
-    defaultConfig {
-        consumerProguardFiles("consumer-rules.pro")
-    }
-
     buildFeatures {
         viewBinding = true
     }
